Added conversion table for BW in RAL layer

+++ Modules/NVMA/NVMA.c
@@ -162,7 +162,7 @@ void NVMA_Get_LR_RX_SF(uint8_t *sf)
  * 
  * @param bw 
  */
-void NVMA_Set_LR_TX_BW(uint32_t bw)
+void NVMA_Set_LR_TX_BW(uint8_t bw)
 {   
     xSemaphoreTake(xEepromMutex, portMAX_DELAY);
     HAL_FLASHEx_DATAEEPROM_Unlock();
@@ -176,7 +176,7 @@ void NVMA_Set_LR_TX_BW(uint32_t bw)
  * 
  * @param bw 
  */
-void NVMA_Get_LR_TX_BW(uint32_t *bw)
+void NVMA_Get_LR_TX_BW(uint8_t *bw)
 {   
     xSemaphoreTake(xEepromMutex, portMAX_DELAY);
     *bw = *((uint32_t *)EE_ADDR_LR_TX_BW);
@@ -188,7 +188,7 @@ void NVMA_Get_LR_TX_BW(uint32_t *bw)
  * 
  * @param bw 
  */
-void NVMA_Set_LR_RX_BW(uint32_t bw)
+void NVMA_Set_LR_RX_BW(uint8_t bw)
 {   
     xSemaphoreTake(xEepromMutex, portMAX_DELAY);
     HAL_FLASHEx_DATAEEPROM_Unlock();
@@ -202,7 +202,7 @@ void NVMA_Set_LR_RX_BW(uint32_t bw)
  * 
  * @param bw 
  */
-void NVMA_Get_LR_RX_BW(uint32_t *bw)
+void NVMA_Get_LR_RX_BW(uint8_t *bw)
 {   
     xSemaphoreTake(xEepromMutex, portMAX_DELAY);
     *bw = *((uint32_t *)EE_ADDR_LR_RX_BW);
